+++ plugins/ucc_chat/lib/ucc_chat_web/views/master_view.ex
@@ -163,4 +163,5 @@ defmodule UccChatWeb.MasterView do
       _ -> ""
     end
   end
+
 end